chore: rename context folder, add todo's
