+++ src/components/themeToggler.tsx
@@ -1,11 +1,13 @@
 /* eslint-disable jsx-a11y/label-has-associated-control */
 import clsx from "clsx";
-import { useThemeContext } from "../services/theme/themeContext";
+import { useThemeContext } from "../context/themeContext";
 import Switch from "./ui/switch";
 
 interface Props {
   className?: string;
 }
+
+// TODO: add tests
 
 const ThemeToggler: React.FC<Props> = ({ className }) => {
   const { currentTheme, toggleTheme } = useThemeContext();
